refactor: move logic to utils and add try catch
